hhk 28th commit

+++ road_following/utility.py
@@ -85,12 +85,12 @@ def preprocess(image, mode, device = "cuda"):
     if mode == "train":
         image = transforms.functional.to_tensor(image)
         image = transforms.functional.normalize(image, [0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
-        return image
+        return image.half()
     if mode == "test":
         image = transforms.functional.to_tensor(image).to(device)
         image = transforms.functional.normalize(image, [0.485, 0.456, 0.406], [0.229, 0.224, 0.225])
         image = image[None, ...]
-        return image
+        return image.half()
 
 
 def dominant_gradient(image):
